requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
pillow==10.0.1
pytesseract==0.3.10
numpy==1.24.3
python-multipart==0.0.6

